Initial documentation, cleanup

+++ src/wp-fragment-html-cache.php
@@ -6,24 +6,25 @@
  * that is bundled with this package in the file license.txt.
  * It is also available through the world-wide-web at this URL:
  * http://www.gnu.org/licenses/gpl-2.0.html
+ *
  * If you did not receive a copy of the license and are unable to
  * obtain it through the world-wide-web, please send an email
  * to info@mindsize.me so we can send you a copy immediately.
  *
  * @package   Mindsize/WP_Fragment_Cache
  * @author    Mindsize
- * @copyright Copyright (c) 2017, Mindsize, LLC.
+ * @copyright Copyright (c) 2017-2021, Mindsize, LLC.
  * @license   http://www.gnu.org/licenses/gpl-2.0.html GNU General Public License v2.0
  */
 
-if( ! defined( 'ABSPATH' ) ) {
+if ( ! defined( 'ABSPATH' ) ) {
 	exit;
 }
 
 /**
  * If the class already exists, no need to redeclare it.
  */
-if( class_exists( 'WP_Fragment_HTML_Cache' ) ) {
+if ( class_exists( 'WP_Fragment_HTML_Cache' ) ) {
 	return;
 }
 
@@ -43,8 +44,10 @@ abstract class WP_Fragment_HTML_Cache extends WP_Fragment_Cache {
 
 	/**
 	 * Get cached data from HTML file and return as a string.
+	 * 
+	 * @todo consider using wp_remote_get().
 	 *
-	 * @param $conditions
+	 * @param array $conditions Array of conditions.
 	 *
 	 * @return string|bool
 	 */
@@ -57,9 +60,10 @@ abstract class WP_Fragment_HTML_Cache extends WP_Fragment_Cache {
 	/**
 	 * Store cached data into HTML file.
 	 *
-	 * @param $data
+	 * @param string $output     Input to be stored (assumed HTML).
+	 * @param array  $conditions Array of conditions.
 	 *
-	 * @return bool
+	 * @return bool If the safe was successful.
 	 */
 	public function set_cache_data( $output, $conditions ) {
 		$bytes  = 0;
@@ -69,14 +73,24 @@ abstract class WP_Fragment_HTML_Cache extends WP_Fragment_Cache {
 
 		$this->ensure_directory_exists( dirname( $file ) );
 
-		if( $cache_file = @fopen( $file, 'w' ) ) {
+		// @todo need to do more to validate $cache_file.
+		// @todo use WP_Filesystem methods.
+		$cache_file = @fopen( $file, 'w' );
+		if ( $cache_file ) {
 			$bytes  = fwrite( $cache_file, $output );
 			$closed = fclose( $cache_file );
 		}
 
-		return 0 < $bytes && true == $closed;
+		return ( 0 < $bytes ) && ( true === $closed );
 	}
 
+	/**
+	 * Fetch the path of the cache file.
+	 *
+	 * @param array $conditions Array of conditions.
+	 *
+	 * @return string The file path.
+	 */
 	public function get_cache_file_path( $conditions ) {
 		/**
 		 * HTML file name will be generated based on passed conditions. Allow for customizing these conditions further.
@@ -93,15 +107,26 @@ abstract class WP_Fragment_HTML_Cache extends WP_Fragment_Cache {
 
 		$file_base = apply_filters( $this->get_hook_name( 'file_base' ), trailingslashit( $this->get_cache_path() ), $file_conditions, $this );
 		$file_name = apply_filters( $this->get_hook_name( 'file_name' ), md5( json_encode( $file_conditions ) ), $file_conditions, $this );
-		$file_path = apply_filters( $this->get_hook_name( 'file_path' ), trailingslashit( $file_base ) . $file_name
-		                                                                 . '.html', $file_base, $file_name, $file_conditions, $this );
+		$file_path = apply_filters(
+			$this->get_hook_name( 'file_path' ),
+			trailingslashit( $file_base ) . $file_name . '.html',
+			$file_base,
+			$file_name,
+			$file_conditions,
+			$this
+		);
+
+		// @todo need further validation, such as file_exists().
+		if ( empty( $file_path ) || ! is_string( $file_path ) ) {
+			return '';
+		}
 
 		return $file_path;
 	}
 
 	/**
 	 * Get the name of the directory for the cache. sanitize_title is run outside of the filter, all contents will be
-	 * sanitiezd.
+	 * sanitized.
 	 *
 	 * @return string
 	 */
@@ -112,54 +137,71 @@ abstract class WP_Fragment_HTML_Cache extends WP_Fragment_Cache {
 	/**
 	 * Get the path to the cache directory, plus any potentially added URI.
 	 *
+	 * @param string $append  Optional. String to append to the cache path.
+	 *
 	 * @return string
 	 */
-	public function get_cache_path( $extra = null ) {
-		$path = trailingslashit( apply_filters( $this->get_hook_name( 'path' ), WP_CONTENT_DIR . '/cache/'
-		                                                                        . $this->get_cache_dir() ) );
+	public function get_cache_path( $append = null ) {
+		$path = trailingslashit(
+			apply_filters(
+				$this->get_hook_name( 'path' ),
+				WP_CONTENT_DIR . '/cache/' . $this->get_cache_dir()
+			)
+		);
 
-		/**
-		 * Add any extra path that was passed
-		 */
-		$path .= $extra;
+		// Add any extra path that was passed.
+		$path .= $append;
 
 		return trailingslashit( $path );
 	}
 
 	/**
 	 * Ensures the cache directory exists, and that the cache is clean.
+	 *
+	 * @param string $append  Optional. String to append to the cache path.
 	 */
-	public function clear_cache( $product_id = null ) {
-		$path = $this->get_cache_path( $product_id );
-
+	public function clear_cache( $append = null ) {
+		$path = $this->get_cache_path( $append );
 		$this->ensure_directory_exists( $path );
 		$this->delete_directory_contents( $path );
 	}
 
 	/**
 	 * Ensure that the HTML cache directory exists, and create it if it does not.
+	 *
+	 * @param string $path Optional. The cache path.
 	 */
 	protected function ensure_directory_exists( $path = null ) {
 		$path = ! empty( $path ) ? $path : $this->get_cache_path();
 
-		if( ! is_dir( $path ) ) {
+		if ( ! is_dir( $path ) ) {
 			wp_mkdir_p( $path );
 		}
 	}
 
 	/**
-	 * Delete an entire directory, does call itself recursively.
+	 * Delete an entire directory. Calls itself recursively.
+	 *
+	 * @param string $dir The directory path to be emptied.
 	 */
 	protected function delete_directory_contents( $dir ) {
-		foreach(
-			array_diff( scandir( $dir ), array(
-				'.',
-				'..'
-			) ) as $file
-		) {
+
+		if ( ! is_string( $dir ) ) {
+			return;
+		}
+
+		// Func scandir can return false.
+		$file_array = scandir( $dir );
+		if ( empty( $file_array ) || ! is_array( $file_array ) ) {
+			return;
+		}
+
+		$files = array_diff( $file_array, array( '.', '..' ) );
+
+		foreach ( $file_array as $file ) {
 			$object = wp_normalize_path( trailingslashit( $dir ) . $file );
 
-			if( is_dir( $object ) ) {
+			if ( is_dir( $object ) ) {
 				$this->delete_directory_contents( $object );
 			} else {
 				unlink( $object );
